refactor(platform): split platform/api/serializers.py into domain modules
Closes #645

Splits the 1462-line flat serializers.py into a package with domain-based modules:
- common.py: shared types (InstructorResponse, LearnerResponse, etc.) to prevent circular imports
- lessons.py, quizzes.py, assignments.py, courses.py, learners.py
- organisations.py, oauth.py, newsletters.py, misc.py
- __init__.py re-exports all 60+ classes for backwards compatibility
